# ⚡ Optimize decoding with less backtracking