Incorporates two following components:

- Less backtracking by avoiding peeking when it isn't necessary
- Returning `byte` instead of `byte?` for `ReadByte()` used for peeking.

## Decoding benchmark

Tested with data size of 10,000. Nested means one additional layer. For example, nested list would be a `List` of size `SetSize` with each element also being a `List` of size `SetSize`. Endpoints were randomized to be either `Binary` or `Text`.

Memory save is mostly from #91. Improves speed by roughly 20% on `List`s and `Dictionary`s.

### 0.12.0

|           Method | SetSize | WordSize |        Mean |     Error |    StdDev |  Allocated |
|----------------- |-------- |--------- |------------:|----------:|----------:|-----------:|
|       DecodeList |      10 |       10 |    16.33 ms |  0.939 ms |  1.005 ms |   33.26 MB |
|       DecodeDict |      10 |       10 |    53.58 ms |  3.509 ms |  4.041 ms |   73.69 MB |
| DecodeNestedList |      10 |       10 |   163.18 ms |  8.270 ms |  9.524 ms |  336.54 MB |
| DecodeNestedDict |      10 |       10 |   574.93 ms | 24.271 ms | 27.951 ms |  781.33 MB |

### Updated

|           Method | SetSize | WordSize |        Mean |     Error |    StdDev |  Allocated |
|----------------- |-------- |--------- |------------:|----------:|----------:|-----------:|
|       DecodeList |      10 |       10 |    11.97 ms |  0.747 ms |  0.830 ms |   21.52 MB |
|       DecodeDict |      10 |       10 |    42.57 ms |  4.061 ms |  4.514 ms |   56.82 MB |
| DecodeNestedList |      10 |       10 |   123.50 ms | 10.638 ms | 11.824 ms |  208.04 MB |
| DecodeNestedDict |      10 |       10 |   439.78 ms | 23.726 ms | 27.323 ms |  596.64 MB |